Simple solution to needing the global SASS context
Many of our per-component styles require the global SASS context. In
particular the helpers and the settings (both ours and bootstraps).  Each time
webpack requires a stylesheet it is compiled in isolation to any other styles,
and must import any dependencies it has.

To avoid importing bootstrap multiple times I've moved the per-component
requires to a single monolithic stylesheet. This makes rebuilds much faster,
but is not what we want. Later we should extract just the settings and helpes
that we do want in such a way that we can make use of those when and where we
need them.