Adapt authorization so function passed in
Pass in the authorization function to `authorize` now, so can authorize
in multiple ways using this. Also removed stuff to do with passing the
component to `authorize` as always using this with the authorized
component as a child element for the moment; can always add this back in
if needed.