Don't redirect to sign in page when signing out
When signing out the loading indicator is displayed.  The develop branch
includes a commit which changed how the swap from "application" page to
loading page was performed.

That commit introduced a fade-in-out animation transition between the
"application" page and the loading page, which causes the "application" page
to remain mounted for a moment longer and to receive updated props.  The
updated props cause the authorisation check to be performed and the redirect
to the sign in page to take place.

This commit prevents that redirection from taking place.  It does this by
maintaining a `signingOut` attribute in the `auth` state, the
`redirectToSignIn` function checks that we are not signing out before
redirecting.