Correct error message customizations
An error message for a particular error code can be customized to provide more
helpful error messages.  It should not be necessary for that customization to
handle all possible failure variations.

Where a failure variation is not handled by the customization we should fall
back to the default message for that error code.  We should not fallback to
"an unknown error".

This commit implements that fallback behaviour.