Correct import to use relative files
Now that this module has been moved to flight-common, this import was
causing an issue as it would attempt to import from the main app instead
of flight-common. Changing this to a relative import should ensure this
keeps working no matter what the directory hierarchy is above the
notification module.