Move info/error generator maps to Redux store
The notification generator maps are now stored in the Redux store; this
prevents an issue with how they were accessed previously where depending
on which module they were imported from (such as the enclosing app or
flight-common internally) different instances of each would be
initialized, which then means the notifications would not be set up when
we attempt to display them.

This is an improvement to the above issue (and moves this aspect of the
state into the store), but is not really in the spirit of Redux as we
still modify these directly rather than via reducers, so time travel
etc.  won't work for these I think.