Don't import markdown
Importing the markdown works fine in development as the files are processed by
webpack complete with its loaders.  In production, the files are first
compiled by babel and the webpack loaders are not used.

The new mechanism to specify the copy, is to pass it as a prop.  The markdown
copy is still available so applications can use webpack to import the markdown
copy first.